+++ aws/transform.go
@@ -80,6 +80,8 @@ func initResource(source interface{}) (*graph.Resource, error) {
 		// Autoscaling
 	case *autoscaling.LaunchConfiguration:
 		res = graph.InitResource(cloud.LaunchConfiguration, awssdk.StringValue(ss.LaunchConfigurationARN))
+	case *autoscaling.Group:
+		res = graph.InitResource(cloud.AutoScalingGroup, awssdk.StringValue(ss.AutoScalingGroupARN))
 	// IAM
 	case *iam.User:
 		res = graph.InitResource(cloud.User, awssdk.StringValue(ss.UserId))